+++ tests/TestCase/Utility/BackupManagerTest.php
@@ -17,7 +17,6 @@ namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\I18n\FrozenTime;
-use Cake\ORM\Entity;
 use Cake\TestSuite\EmailTrait;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
@@ -111,9 +110,9 @@ class BackupManagerTest extends TestCase
 
         //Checks for properties of each backup object
         foreach ($files as $file) {
-            $this->assertInstanceOf(Entity::class, $file);
-            $this->assertGreaterThan(0, $file->get('size'));
-            $this->assertInstanceOf(FrozenTime::class, $file->get('datetime'));
+            $this->assertIsArray($file);
+            $this->assertGreaterThan(0, $file['size']);
+            $this->assertInstanceOf(FrozenTime::class, $file['datetime']);
         }
     }
 
@@ -139,7 +138,11 @@ class BackupManagerTest extends TestCase
         $this->assertSame(['bzip2', 'gzip'], $filesAfterRotate->extract('compression')->toList());
 
         //Gets the difference
-        $diff = array_udiff($initialFiles->toList(), $filesAfterRotate->toList(), fn(Entity $first, Entity $second): int => strcmp($first->get('filename'), $second->get('filename')));
+        $diff = array_udiff(
+            $initialFiles->toList(),
+            $filesAfterRotate->toList(),
+            fn(array $first, array $second): int => strcmp($first['filename'], $second['filename'])
+        );
 
         //Again, only 1 backup was deleted. The difference is the same
         $this->assertCount(1, $diff);