+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * it no longer needs the `me-tools` package. This removes several (useless) dependencies;
 * some, possible changes that prepare it for CakePHP 5 and PHPUnit 10 ([issue #97](https://github.com/mirko-pagliai/cakephp-database-backup/issues/97));
 * little fixes. Fixed some deprecations for CakePHP 4.5 ([issue #97](https://github.com/mirko-pagliai/cakephp-database-backup/issues/97));
-* improved `BackuManager::index()` method;
+* improved `BackuManager::index()` method, also regarding the correct files sorting. This also solves a small bug for
+  the `rotate()` method (which precisely affects `index()`);
 * some testing methods that have been missing for a long time have been added;
 * the `BackupTrait::getDriverName()` method can no longer be static;
 * removed (old and useless) `BaseCommandTestCase` class;